pyproject.toml
[project]
name = "datafed-script"
version = "0.1.0"
description = "TBD"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "datafed>=3.0.3",
  "mypy>=1.15.0",
  "protobuf>=6.30.1",
  "pyright>=1.1.396",
  "python-dotenv>=1.0.1",
  "ruff>=0.11.0",
]
